+++ lib/credo/check/readability/string_sigils.ex
@@ -31,6 +31,9 @@ defmodule Credo.Check.Readability.StringSigils do
   ]
   @quote_codepoint 34
 
+  alias Credo.SourceFile
+  alias Credo.Code.Heredocs
+
   use Credo.Check, base_priority: :low
 
   @doc false
@@ -39,18 +42,29 @@ defmodule Credo.Check.Readability.StringSigils do
 
     maximum_allowed_quotes = Params.get(params, :maximum_allowed_quotes, @default_params)
 
-    Credo.Code.prewalk(
-      source_file,
-      &traverse(&1, &2, issue_meta, maximum_allowed_quotes)
-    )
+    case remove_heredocs_and_convert_to_ast(source_file) do
+      {:ok, ast} ->
+        Credo.Code.prewalk(ast, &traverse(&1, &2, issue_meta, maximum_allowed_quotes))
+
+      {:error, errors} ->
+        IO.warn("Unexpected error while parsing #{source_file.filename}: #{inspect(errors)}")
+        []
+    end
   end
 
-  def traverse(
-        {maybe_sigil, meta, [str | rest_ast]} = ast,
-        issues,
-        issue_meta,
-        maximum_allowed_quotes
-      ) do
+  defp remove_heredocs_and_convert_to_ast(source_file) do
+    source_file
+    |> SourceFile.source()
+    |> Heredocs.replace_with_spaces()
+    |> Credo.Code.ast()
+  end
+
+  defp traverse(
+         {maybe_sigil, meta, [str | rest_ast]} = ast,
+         issues,
+         issue_meta,
+         maximum_allowed_quotes
+       ) do
     line_no = meta[:line]
 
     cond do
@@ -74,7 +88,7 @@ defmodule Credo.Check.Readability.StringSigils do
     end
   end
 
-  def traverse(ast, issues, _issue_meta, _maximum_allowed_quotes) do
+  defp traverse(ast, issues, _issue_meta, _maximum_allowed_quotes) do
     {ast, issues}
   end
 
@@ -93,18 +107,11 @@ defmodule Credo.Check.Readability.StringSigils do
          issue_meta,
          line_no
        ) do
-    if !is_heredoc(issue_meta, line_no) && too_many_quotes?(string, maximum_allowed_quotes) do
+    if too_many_quotes?(string, maximum_allowed_quotes) do
       [issue_for(issue_meta, line_no, string, maximum_allowed_quotes) | issues]
     else
       issues
     end
-  end
-
-  defp is_heredoc({_, source_file, _}, line_no) do
-    lines = SourceFile.lines(source_file)
-    {_, line} = Enum.find(lines, fn {n, _} -> n == line_no end)
-
-    Regex.match?(~r/("""|''')$/, line)
   end
 
   defp too_many_quotes?(string, limit) do