+++ lib/credo/code/name.ex
@@ -77,8 +77,10 @@ defmodule Credo.Code.Name do
     "@#{name}"
   end
 
-  def full({:unquote, _, [{name, _, nil}]}) when is_atom(name) do
-    "unquote(#{name})"
+  def full({name, _, arguments}) when is_atom(name) and is_list(arguments) do
+    arg_list = arguments |> Enum.map(&full/1) |> Enum.join(", ")
+
+    "#{full(name)}(#{arg_list})"
   end
 
   def parts_count(module_name) do