+++ lib/credo/code/strings.ex
@@ -63,7 +63,7 @@ defmodule Credo.Code.Strings do
   end
 
   defp parse_code(<<"\'\'\'"::utf8, t::binary>>, acc, replacement) do
-    parse_heredoc(t, acc <> ~s("""), replacement)
+    parse_heredoc(t, acc <> ~s('''), replacement)
   end
 
   defp parse_code(<<"\""::utf8, t::binary>>, acc, replacement) do
@@ -183,6 +183,10 @@ defmodule Credo.Code.Strings do
 
   defp parse_heredoc(<<"\"\"\""::utf8, t::binary>>, acc, replacement) do
     parse_code(t, acc <> ~s("""), replacement)
+  end
+
+  defp parse_heredoc(<<"\'\'\'"::utf8, t::binary>>, acc, replacement) do
+    parse_code(t, acc <> ~s('''), replacement)
   end
 
   defp parse_heredoc(<<"\n"::utf8, t::binary>>, acc, replacement) do